refactor(Breadcrumbs): remove the interface to index file
