--- src/components/navigation/Breadcrumbs/interfaces/Breadcrumbs.Item.interface.ts
@@ -1,6 +1,0 @@
-export interface IBreadcrumbItem {
-  path: string;
-  label: string;
-  id: string;
-  isActive: boolean;
-}
