--- src/components/navigation/Breadcrumbs/interfaces/Breadcrumbs.interface.ts
@@ -1,4 +1,0 @@
-import { IBreadcrumbItem } from "./Breadcrumbs.Item.interface";
-export interface IBreadcrumbsProps {
-  crumbs: IBreadcrumbItem[];
-}
