+++ src/components/navigation/Breadcrumbs/stories/Breadcrumbs.Default.stories.tsx
@@ -1,7 +1,7 @@
 import { Breadcrumbs } from "../index";
 import { BrowserRouter } from "react-router-dom";
 import { parameters, crumbs } from "./props";
-import { IBreadcrumbsProps } from "../interfaces/Breadcrumbs.interface";
+import { IBreadcrumbsProps } from "..";
 
 const story = {
   title: "navigation/Breadcrumbs",
